python_agents/requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langchain-core>=0.3.72
langchain-ollama==0.3.6
pydantic>=2.7.4
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
httpx>=0.27.0
aiohttp>=3.8.0
aiofiles>=23.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv==1.0.0
jira==3.5.2
chromadb==0.4.22